requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
django-celery-beat==2.5.0
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.0.1
django-extensions==3.2.3
pytest-django==4.5.2
pytest==7.4.2
factory-boy==3.3.0
freezegun==1.2.2